[S1-Ilia] add check database and fix dto

+++ api/src/main/java/com/knits/ammolite/model/asset/Depreciation.java
@@ -27,7 +27,7 @@ public class Depreciation {
     @Column
     private DepreciationMethodType depreciationMethod;
     @Column
-    private Integer salvation;
+    private BigDecimal salvation;
 
 
 }
